AlphaBetaEngine/requirements.txt
python-chess==0.23.10
tensorflow-gpu==1.3.0
keras==2.0.8
profilehooks
numpy
pyperclip
h5py==2.8.0
